feat: 운동 종료 시 회원의 운동 정보 업데이트 기능 구현

+++ src/main/java/clofi/runningplanet/member/dto/response/ProfileResponse.java
@@ -15,9 +15,9 @@ public record ProfileResponse(
 
 	AvgPace avgPace,
 
-	int avgDistance,
+	double avgDistance,
 
-	int totalDistance,
+	double totalDistance,
 
 	String myCrew
 ) {
